% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impute_spring.R
\name{impute_spring}
\alias{impute_spring}
\title{Impute prior spring RIT from current fall}
\usage{
impute_spring(input, subject = "reading", na_replacement = -1)
}
\arguments{
\item{input}{data passed on as \code{newdata} to \code{\link{predict}}.
requirese the following predictors: \code{TestRITScore.x}, \code{Goal1RitScore.x},
\code{Goal2RitScore.x}, \code{Goal3RitScore.x}, \code{Goal4RitScore.x},
\code{TestDurationMinutes.x}, \code{PercentCorrect.x}, and \code{Grade}}

\item{subject}{character vector indicating \code{"reading"} or \code{"math"}}

\item{na_replacment}{numerical value used to replace \code{NA}s in data data.}
}
\description{
Models require 5 predictors \code{TestRITScore.x}, \code{Goal1RitScore.x},
\code{Goal2RitScore.x}, \code{Goal3RitScore.x}, \code{Goal4RitScore.x},
\code{TestDurationMinutes.x}, \code{PercentCorrect.x}, and \code{Grade},
as well as a subject indicator.
}
\examples{
require(mapvisuals)
require(dplyr)
data(nweamap)

map_mv <- nweamap \%>\% filter(MeasurementScale=="Reading") \%>\%
   mapvizer

test_data<-map_mv \%>\% head(10)

impute_spring(test_data, subject="reading")
}

